Move #getPermissionChecker
