+++ api/src/main/java/com/velocitypowered/api/command/CommandSource.java
@@ -8,15 +8,12 @@
 package com.velocitypowered.api.command;
 
 import com.velocitypowered.api.permission.PermissionSubject;
-import com.velocitypowered.api.permission.Tristate;
 import net.kyori.adventure.audience.Audience;
 import net.kyori.adventure.audience.MessageType;
 import net.kyori.adventure.identity.Identified;
 import net.kyori.adventure.identity.Identity;
-import net.kyori.adventure.permission.PermissionChecker;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.serializer.legacytext3.LegacyText3ComponentSerializer;
-import net.kyori.adventure.util.TriState;
 import org.checkerframework.checker.nullness.qual.NonNull;
 
 /**
@@ -38,25 +35,5 @@ public interface CommandSource extends Audience, PermissionSubject {
   default void sendMessage(@NonNull Identity identity, @NonNull Component message,
                            @NonNull MessageType type) {
     this.sendMessage(LegacyText3ComponentSerializer.get().serialize(message));
-  }
-
-  /**
-   * Gets the permission checker for the invoker.
-   *
-   * @return invoker's permission checker
-   */
-  default PermissionChecker getPermissionChecker() {
-    return permission -> {
-      final Tristate state = getPermissionValue(permission);
-      if (state == Tristate.TRUE) {
-        return TriState.TRUE;
-      } else if (state == Tristate.UNDEFINED) {
-        return TriState.NOT_SET;
-      } else if (state == Tristate.FALSE) {
-        return TriState.FALSE;
-      } else {
-        throw new IllegalArgumentException();
-      }
-    };
   }
 }
