+++ api/src/main/java/com/velocitypowered/api/permission/PermissionSubject.java
@@ -7,6 +7,9 @@
 
 package com.velocitypowered.api.permission;
 
+import net.kyori.adventure.permission.PermissionChecker;
+import net.kyori.adventure.util.TriState;
+
 /**
  * Represents a object that has a set of queryable permissions.
  */
@@ -29,4 +32,24 @@ public interface PermissionSubject {
    * @return the value the permission is set to
    */
   Tristate getPermissionValue(String permission);
+
+  /**
+   * Gets the permission checker for the subject.
+   *
+   * @return subject's permission checker
+   */
+  default PermissionChecker getPermissionChecker() {
+    return permission -> {
+      final Tristate state = getPermissionValue(permission);
+      if (state == Tristate.TRUE) {
+        return TriState.TRUE;
+      } else if (state == Tristate.UNDEFINED) {
+        return TriState.NOT_SET;
+      } else if (state == Tristate.FALSE) {
+        return TriState.FALSE;
+      } else {
+        throw new IllegalArgumentException();
+      }
+    };
+  }
 }
